+++ eaglestitch/stitching_api/module.py
@@ -1,6 +1,7 @@
 import asab
 from .service import StitchingAPIService
 from .stitching_result import StitchingResultSectionWebHandler
+from .stitching_controller import StitchingControllerSectionWebHandler
 import logging
 
 ###
@@ -17,6 +18,7 @@ class StitchingAPIModule(asab.Module):
 		super().__init__(app)
 		self.Service = StitchingAPIService(app)
 		self.StitchingResultSectionWebHandler = StitchingResultSectionWebHandler(app)
+		self.StitchingControllerSectionWebHandler = StitchingControllerSectionWebHandler(app)
 
 	async def initialize(self, app):
 		L.warning("[SYSTEM PREPARATION] Stitching API Module module is loaded.")