+++ eaglestitch/stitching_api/stitching_result/handler.py
@@ -16,7 +16,9 @@ class StitchingResultSectionWebHandler(object):
 		self.stitching_api_svc = app.get_service("eaglestitch.StitchingAPIService")
 
 		app.RESTWebContainer.WebApp.router.add_get('/stitching', self.get_stitching)
-		app.RESTWebContainer.WebApp.router.add_get('/stitching/{stitched_id}', self.get_stitching_by_id)
+		app.RESTWebContainer.WebApp.router.add_get('/stitching/{stitch_id}', self.get_stitching_by_id)
+
+		self.to_url = asab.Config["eaglestitch:rest"].getboolean("to_url")
 
 	async def get_stitching(self, request):
 		status, msg, data = await self.stitching_api_svc.get_stitching_result()
@@ -29,13 +31,14 @@ class StitchingResultSectionWebHandler(object):
 		}, status=status)
 
 	async def get_stitching_by_id(self, request):
-		stitched_id = (request.match_info["stitched_id"]).upper()
-		if stitched_id is None:
-			L.error("`stitched_id` is missing from request.")
+		stitch_id = (request.match_info["stitch_id"]).upper()
+		if stitch_id is None:
+			L.error("`stitch_id` is missing from request.")
 			raise aiohttp.web.HTTPBadRequest()
 
 		status, msg, data = await self.stitching_api_svc.get_stitching_result(
-			stitched_id=stitched_id
+			stitch_id=stitch_id,
+			to_url=self.to_url,
 		)
 
 		# Here we received the barrier and we can respond 'OK'